scripts/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0





